1st Iteration
I wrote this code myself. It was not fully successful. I have not yet added the influence of the parental blood type per person. As of now there is only an allele randomizer incantation on the blood type per person. I still want to incorporate the use of this randomizer function. I will make one more iteration here which should be the final form.